refactor: headunit can transceiver class
