--- Project-3/HeadUnit/canreciever.cpp
@@ -1,6 +1,0 @@
-#include "canreciever.h"
-
-CanReciever::CanReciever()
-{
-
-}
