--- Project-3/HeadUnit/canreciever.h
@@ -1,15 +1,0 @@
-#ifndef CANRECIEVER_H
-#define CANRECIEVER_H
-
-#include <QObject>
-
-// CAN definition
-# define CAN_FRAME_MAX_LEN 8
-
-class CanReciever
-{
-public:
-    CanReciever();
-};
-
-#endif // CANRECIEVER_H
